# src/openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "Library Management System API",
		"version": "1.1.0",
		"description": "REST API for a library management system — books, authors, users, and the borrow/return workflow, with role-based access control (Admin/Librarian/User).\n\nThis document is generated from JSDoc annotations on the route files (`npm run docs:generate`) — do not hand-edit `docs/API.md` or `src/openapi.json`.",
		"license": {
			"name": "ISC"
		}
	},
	"servers": [
		{
			"url": "http://localhost:5080/api",
			"description": "Local development"
		},
		{
			"url": "{baseUrl}/api",
			"description": "Custom",
			"variables": {
				"baseUrl": {
					"default": "http://localhost:5080"
				}
			}
		}
	],
...
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Obtain a token from `POST /auth/register`, `POST /auth/login`, or `POST /auth/setup-admin`, then send it as `Authorization: Bearer <token>`."
			}
		},
		"schemas": {
			"Error": {
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean",
						"example": false
					},
					"message": {
						"type": "string",
						"example": "What went wrong"
					},
					"error": {
						"type": "string",
						"description": "Detail, omitted in production for 500s."
					}
				}
			},
			"ValidationError": {
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean",
						"example": false
					},
					"message": {
						"type": "string",
						"example": "Validation failed"
					},
					"errors": {
						"type": "array",
						"items": {
							"type": "object",
							"properties": {
								"field": {
									"type": "string",
									"example": "email"
								},
								"message": {
									"type": "string",
									"example": "Invalid email format"
								}
							}
						}
					}
				}
			},
			"Pagination": {
				"type": "object",
				"properties": {
					"current_page": {
						"type": "integer",
						"example": 1
					},
					"total_pages": {
						"type": "integer",
						"example": 5
					},
					"total_items": {
						"type": "integer",
						"example": 47
					},
					"items_per_page": {
						"type": "integer",
						"example": 10
					},
					"has_next": {
						"type": "boolean",
						"example": true
					},
					"has_prev": {
						"type": "boolean",
						"example": false
					}
				}
			},
			"User": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"example": 1
					},
					"first_name": {
						"type": "string",
						"example": "John"
					},
					"last_name": {
						"type": "string",
						"example": "Doe"
					},
					"full_name": {
						"type": "string",
						"example": "John Doe"
					},
					"user_name": {
						"type": "string",
						"example": "johndoe"
					},
					"email": {
						"type": "string",
						"format": "email",
						"example": "john@example.com"
					},
					"phone": {
						"type": "string",
						"nullable": true,
						"example": "08012345678"
					},
					"image_url": {
						"type": "string",
						"nullable": true
					},
					"role": {
						"type": "string",
						"enum": ["Admin", "Librarian", "User"],
						"example": "User"
					},
					"is_active": {
						"type": "boolean",
						"example": true
					},
					"email_verified": {
						"type": "boolean",
						"example": false
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
...
			"Author": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"example": 1
					},
					"first_name": {
						"type": "string",
						"example": "George"
					},
					"last_name": {
						"type": "string",
						"example": "Orwell"
					},
					"email": {
						"type": "string",
						"format": "email",
						"example": "orwell@example.com"
					},
					"phone": {
						"type": "string",
						"nullable": true
					},
					"image": {
						"type": "string",
						"nullable": true
					},
					"date_of_birth": {
						"type": "string",
						"format": "date",
						"nullable": true
					},
					"biography": {
						"type": "string",
						"nullable": true
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"Book": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"example": 1
					},
					"isbn": {
						"type": "string",
						"example": "978-0451524935"
					},
					"title": {
						"type": "string",
						"example": "1984"
					},
					"author_id": {
						"type": "integer",
						"example": 1
					},
					"published_date": {
						"type": "string",
						"format": "date",
						"nullable": true
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"cover_image": {
						"type": "string",
						"nullable": true
					},
					"genre": {
						"type": "string",
						"nullable": true,
						"example": "Dystopian Fiction"
					},
					"language": {
						"type": "string",
						"nullable": true,
						"example": "English"
					},
					"pages": {
						"type": "integer",
						"nullable": true,
						"example": 328
					},
					"publisher": {
						"type": "string",
						"nullable": true
					},
					"available_copies": {
						"type": "integer",
						"example": 5
					},
					"total_copies": {
						"type": "integer",
						"example": 5
					},
					"status": {
						"type": "string",
						"enum": ["Available", "Borrowed", "Reserved", "Lost"],
						"example": "Available"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"BorrowRecord": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"example": 10
					},
					"user_id": {
						"type": "integer",
						"example": 1
					},
					"book_id": {
						"type": "integer",
						"example": 1
					},
					"borrow_date": {
						"type": "string",
						"format": "date-time"
					},
					"due_date": {
						"type": "string",
						"format": "date-time"
					},
					"return_date": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"status": {
						"type": "string",
						"enum": ["Borrowed", "Returned", "Overdue"],
						"example": "Borrowed"
					},
					"is_active": {
						"type": "boolean",
						"example": true
					}
				}
			}
		},
		"responses": {
			"Unauthorized": {
				"description": "Missing, invalid, or expired token",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Error"
						},
						"example": {
							"success": false,
							"message": "Access denied. No token provided"
						}
					}
				}
			},
			"Forbidden": {
				"description": "Authenticated, but wrong role or not the resource owner",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Error"
						},
						"example": {
							"success": false,
							"message": "Access denied. Required role(s): Admin. Your role: User"
						}
					}
				}
			},
...
			"ValidationFailed": {
				"description": "Request body failed schema validation",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/ValidationError"
						}
					}
				}
			},
...
		}
	},
	"security": [
		{
			"bearerAuth": []
		}
	],
	"paths": {
...
		"/users": {
			"get": {
				"tags": ["Users"],
				"summary": "List all users",
				"description": "Requires the Admin role.",
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 10
						}
					},
					{
						"name": "search",
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "role",
						"in": "query",
						"schema": {
							"type": "string",
							"enum": ["Admin", "Librarian", "User"]
						}
					},
					{
						"name": "is_active",
						"in": "query",
						"schema": {
							"type": "boolean"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Paginated user list.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"data": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/User"
											}
										},
										"pagination": {
											"$ref": "#/components/schemas/Pagination"
										}
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"403": {
						"$ref": "#/components/responses/Forbidden"
					}
				}
			},
			"post": {
				"tags": ["Users"],
				"summary": "Create a user",
				"description": "Requires the Admin role. Can create any role, unlike self-registration.",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": [
									"first_name",
									"last_name",
									"user_name",
									"email",
									"password"
								],
								"properties": {
									"first_name": {
										"type": "string",
										"example": "Jane"
									},
									"last_name": {
										"type": "string",
										"example": "Smith"
									},
									"user_name": {
										"type": "string",
										"minLength": 3,
										"example": "janesmith"
									},
									"email": {
										"type": "string",
										"format": "email"
									},
									"phone": {
										"type": "string",
										"nullable": true
									},
									"password": {
										"type": "string",
										"example": "Password123"
									},
									"image_url": {
										"type": "string",
										"nullable": true
									},
									"role": {
										"type": "string",
										"enum": ["Admin", "Librarian", "User"],
										"default": "User"
									},
									"is_active": {
										"type": "boolean",
										"default": true
									},
									"email_verified": {
										"type": "boolean",
										"default": false
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "User created.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"data": {
											"$ref": "#/components/schemas/User"
										}
									}
								}
							}
						}
					},
					"400": {
						"$ref": "#/components/responses/ValidationFailed"
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"403": {
						"$ref": "#/components/responses/Forbidden"
					},
					"409": {
						"description": "Email or username already taken.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			}
		},
...
		"/books": {
			"get": {
				"tags": ["Books"],
				"summary": "List books",
				"description": "Responses are cached for `BOOKS_CACHE_TTL_SECONDS` (default 30s), keyed by the full query string, and invalidated on any create/update/delete.\n",
				"security": [],
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 10
						}
					},
					{
						"name": "search",
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "author_id",
						"in": "query",
						"schema": {
							"type": "integer"
						}
					},
					{
						"name": "genre",
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "status",
						"in": "query",
						"schema": {
							"type": "string",
							"enum": ["Available", "Borrowed", "Reserved", "Lost"]
						}
					}
				],
				"responses": {
					"200": {
						"description": "Paginated book list.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"data": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/Book"
											}
										},
										"pagination": {
											"$ref": "#/components/schemas/Pagination"
										}
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": ["Books"],
				"summary": "Create a book",
				"description": "Requires the Admin or Librarian role.",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": ["isbn", "title", "author_id"],
								"properties": {
									"isbn": {
										"type": "string",
										"example": "978-0451524935"
									},
									"title": {
										"type": "string",
										"example": "1984"
									},
									"author_id": {
										"type": "integer",
										"example": 1
									},
									"published_date": {
										"type": "string",
										"format": "date",
										"nullable": true
									},
									"description": {
										"type": "string",
										"nullable": true
									},
									"cover_image": {
										"type": "string",
										"nullable": true
									},
									"genre": {
										"type": "string",
										"nullable": true,
										"example": "Dystopian Fiction"
									},
									"language": {
										"type": "string",
										"nullable": true,
										"example": "English"
									},
									"pages": {
										"type": "integer",
										"nullable": true,
										"example": 328
									},
									"publisher": {
										"type": "string",
										"nullable": true
									},
									"available_copies": {
										"type": "integer",
										"default": 0
									},
									"total_copies": {
										"type": "integer",
										"default": 0
									},
									"status": {
										"type": "string",
										"enum": ["Available", "Borrowed", "Reserved", "Lost"],
										"default": "Available"
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Book created.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"data": {
											"$ref": "#/components/schemas/Book"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Validation failure, or available_copies exceeds total_copies.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationError"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"403": {
						"$ref": "#/components/responses/Forbidden"
					},
					"404": {
						"description": "author_id does not exist.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"409": {
						"description": "A book with this ISBN already exists.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			}
		},
...
		"/borrow-records": {
			"get": {
				"tags": ["Borrow Records"],
				"summary": "List all borrow records",
				"description": "Requires the Admin or Librarian role.",
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 10
						}
					},
					{
						"name": "user_id",
						"in": "query",
						"schema": {
							"type": "integer"
						}
					},
					{
						"name": "book_id",
						"in": "query",
						"schema": {
							"type": "integer"
						}
					},
					{
						"name": "status",
						"in": "query",
						"schema": {
							"type": "string",
							"enum": ["Borrowed", "Returned", "Overdue"]
						}
					},
					{
						"name": "overdue_only",
						"in": "query",
						"schema": {
							"type": "boolean",
							"default": false
						}
					}
				],
				"responses": {
					"200": {
						"description": "Paginated borrow records.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"data": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/BorrowRecord"
											}
										},
										"pagination": {
											"$ref": "#/components/schemas/Pagination"
										}
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"403": {
						"$ref": "#/components/responses/Forbidden"
					}
				}
			}
		},
...
		"/authors": {
			"get": {
				"tags": ["Authors"],
				"summary": "List authors",
				"description": "Responses are cached for `AUTHORS_CACHE_TTL_SECONDS` (default 30s), keyed by the full query string, and invalidated on any create/update/delete.\n",
				"security": [],
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 10
						}
					},
					{
						"name": "search",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"description": "Matches first_name, last_name, or email."
					},
					{
						"name": "sort_by",
						"in": "query",
						"schema": {
							"type": "string",
							"default": "created_at"
						}
					},
					{
						"name": "order",
						"in": "query",
						"schema": {
							"type": "string",
							"enum": ["asc", "desc"],
							"default": "desc"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Paginated author list.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"data": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/Author"
											}
										},
										"pagination": {
											"$ref": "#/components/schemas/Pagination"
										}
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": ["Authors"],
				"summary": "Create an author",
				"description": "Requires the Admin or Librarian role.",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": ["first_name", "last_name", "email"],
								"properties": {
									"first_name": {
										"type": "string",
										"example": "George"
									},
									"last_name": {
										"type": "string",
										"example": "Orwell"
									},
									"email": {
										"type": "string",
										"format": "email",
										"example": "orwell@example.com"
									},
									"date_of_birth": {
										"type": "string",
										"format": "date",
										"nullable": true
									},
									"biography": {
										"type": "string",
										"nullable": true
									},
									"phone": {
										"type": "string",
										"nullable": true
									},
									"image": {
										"type": "string",
										"nullable": true
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Author created.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"data": {
											"$ref": "#/components/schemas/Author"
										}
									}
								}
							}
						}
					},
					"400": {
						"$ref": "#/components/responses/ValidationFailed"
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"403": {
						"$ref": "#/components/responses/Forbidden"
					},
					"409": {
						"description": "An author with this email already exists.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			}
		},
...
	}
}